+++ src/api/MovesService.ts
@@ -1,4 +1,4 @@
-import { getRequest } from '@/utils/requestFunctions';
+import { getRequest, postRequest } from '@/utils/requestFunctions';
 
 export const getMovesListData = async (
   page: number,
@@ -25,5 +25,20 @@ export const getMovesListData = async (
   } catch (error) {
     console.error('Get Move List Fetch Error:', error);
     throw error;
+  }
+};
+
+export const postQuotation = async (serviceType: string, date: string, fromAddress: string, toAddress: string) => {
+  try {
+    const requestBody = {
+      serviceType,
+      date,
+      fromAddress,
+      toAddress,
+    };
+    const data = await postRequest('/moves', requestBody);
+    return data;
+  } catch (err) {
+    console.error('Post quotation error: ', err);
   }
 };
