--- src/api/QuotationService.ts
@@ -1,16 +1,0 @@
-import { postRequest } from '@/utils/requestFunctions';
-
-export const postQuotation = async (serviceType: string, date: string, fromAddress: string, toAddress: string) => {
-  try {
-    const requestBody = {
-      serviceType,
-      date,
-      fromAddress,
-      toAddress,
-    };
-    const data = await postRequest('/moves', requestBody);
-    return data;
-  } catch (err) {
-    console.error('Post quotation error: ', err);
-  }
-};
